requirements.txt
aiogram>=3.7.0
python-dotenv>=1.0.1
tinkoff-investments>=0.2.0b100
pandas>=2.2.2
